Adding entities and repositories for player summary tables

+++ src/main/java/com/varunsrini/cricbase/cricketdatabase/entity/data/BowlingAnalysis.java
@@ -24,7 +24,7 @@ public class BowlingAnalysis {
 
     @Column(name = "OversBowled")
     @JsonProperty("oversBowled")
-    private int oversBowled;
+    private float oversBowled;
 
     @Column(name = "RunsConceded")
     @JsonProperty("runsConceded")
